+++ src/components/Icons/IconSvgs.tsx
@@ -10,6 +10,9 @@ import arrow from "../../../icons/svg/arrow.svg";
 import check from "../../../icons/svg/check.svg";
 import clock from "../../../icons/svg/clock.svg";
 import close from "../../../icons/svg/close.svg";
+import download from "../../../icons/svg/download.svg";
+import errorFilled from "../../../icons/svg/error-filled.svg";
+import errorOutline from "../../../icons/svg/error-outline.svg";
 import fileTypeAudio from "../../../icons/svg/file-type-audio.svg";
 import fileTypeDoc from "../../../icons/svg/file-type-doc.svg";
 import fileTypeGenericDoc from "../../../icons/svg/file-type-generic-doc.svg";
@@ -17,13 +20,15 @@ import fileTypeImage from "../../../icons/svg/file-type-image.svg";
 import fileTypePdf from "../../../icons/svg/file-type-pdf.svg";
 import fileTypeSpreadsheet from "../../../icons/svg/file-type-spreadsheet.svg";
 import fileTypeVideo from "../../../icons/svg/file-type-video.svg";
-import download from "../../../icons/svg/download.svg";
-import errorFilled from "../../../icons/svg/error-filled.svg";
-import errorOutline from "../../../icons/svg/error-outline.svg";
 import headset from "../../../icons/svg/headset.svg";
 import minus from "../../../icons/svg/minus.svg";
 import plus from "../../../icons/svg/plus.svg";
 import search from "../../../icons/svg/search.svg";
+import socialFacebook from "../../../icons/svg/social-facebook.svg";
+import socialInstagram from "../../../icons/svg/social-instagram.svg";
+import socialTumblr from "../../../icons/svg/social-tumblr.svg";
+import socialTwitter from "../../../icons/svg/social-twitter.svg";
+import socialYoutube from "../../../icons/svg/social-youtube.svg";
 import speakerNotes from "../../../icons/svg/speaker-notes.svg";
 import utilityAccountFilled from "../../../icons/svg/utility-account-filled.svg";
 import utilityAccountUnfilled from "../../../icons/svg/utility-account-unfilled.svg";
@@ -56,6 +61,11 @@ export default {
   minus,
   plus,
   search,
+  socialFacebook,
+  socialInstagram,
+  socialTumblr,
+  socialTwitter,
+  socialYoutube,
   speakerNotes,
   utilityAccountFilled,
   utilityAccountUnfilled,